requirements.txt
pandas
yfinance
pyyaml
